// projects/miragon-example-webapp/src/api/models/update-project-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * Data to update an existing io.miragon.example.base.project
 * @export
 * @interface UpdateProjectTO
 */
export interface UpdateProjectTO {
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectTO
     */
    customer: any;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectTO
     */
    address: any;
}